adk_adapter/pyproject.toml
[project]
name = "adk-adapter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "ag-ui-adk>=0.4.0",
    "dotenv>=0.9.9",
    "fastapi>=0.128.0",
    "google-adk>=1.14.1",
    "uvicorn>=0.40.0",
]
